Atualização do desempenho dos candidatos
