+++ CandidatoTableModel.java
@@ -5,7 +5,7 @@ import javax.swing.table.AbstractTableModel;
 
 public class CandidatoTableModel extends AbstractTableModel {
     private List<Candidato> candidatos = new ArrayList<>();
-    private String[] colunas = new String[]{"Id", 
+    private String[] colunas = new String[]{"idCandidato", 
                                             "Nome",
                                             "Numero",
                                             "Partido"};
