+++ Desempenho.java
@@ -10,37 +10,49 @@ public class Desempenho {
     public String getNomeCandidato() {
         return nomeCandidato;
     }
+    
     public void setNomeCandidato(String nomeCandidato) {
         this.nomeCandidato = nomeCandidato;
     }
+
     public String getSiglaPartido() {
         return siglaPartido;
     }
+
     public void setSiglaPartido(String siglaPartido) {
         this.siglaPartido = siglaPartido;
     }
+
     public int getIntencaoVotos() {
         return intencaoVotos;
     }
+
     public void setIntencaoVotos(int intencaoVotos) {
         this.intencaoVotos = intencaoVotos;
     }
+
     public String getFontePesquisa() {
         return fontePesquisa;
     }
+
     public void setFontePesquisa(String fontePesquisa) {
         this.fontePesquisa = fontePesquisa;
     }
+
     public int getMesPesquisa() {
         return mesPesquisa;
     }
+
     public void setMesPesquisa(int mesPesquisa) {
         this.mesPesquisa = mesPesquisa;
     }
+
     public int getAnoPesquisa() {
         return anoPesquisa;
     }
+
     public void setAnoPesquisa(int anoPesquisa) {
         this.anoPesquisa = anoPesquisa;
     }
-}   
+
+}
